Add files via upload
A program that will ask the user for the type of ingredients they would like and the weight of the ingredients in grams
until "no" is entered into the program when asked if there are more ingredients. It will then provide the number of ingredients
entered and the total weight in grams of the ingredients.